Move getNMical into ICalGenerator.ts
