+++ app/logic/Calendar/Event.ts
@@ -1,7 +1,7 @@
 import type { Calendar } from "./Calendar";
 import type { Participant } from "./Participant";
 import type { RecurrenceRule } from "./RecurrenceRule";
-import { ResponseType, type Responses, ParticipationStatus, type iCalMethod } from "./Invitation";
+import { ResponseType, type Responses } from "./Invitation";
 import { appGlobal } from "../app";
 import { Observable, notifyChangedAccessor, notifyChangedProperty } from "../util/Observable";
 import { Lock } from "../util/Lock";
@@ -201,61 +201,6 @@ export class Event extends Observable {
     await accounts[0].sendInvitationResponse(this, response);
   }
 
-  /* Returns an icalEvent object suitable for NodeMailer */
-  getNMical(method?: iCalMethod): { method: iCalMethod, content: string } | null {
-    if (!method) {
-      return null;
-    }
-    /* We have to special-case RRULE as it contains ";"s
-     * which must not be escaped as normal text values would */
-    const lines: (string | string[])[] = [];
-    lines.push(["BEGIN", "VCALENDAR"]);
-    lines.push(["METHOD", method]);
-    lines.push(["VERSION", "2.0"]);
-    lines.push(["PRODID", "-//Beonex//Parula//EN"]);
-    lines.push(["BEGIN", "VEVENT"]);
-    lines.push(["DTSTAMP", date2ical(new Date(), false)]);
-    lines.push(["UID", this.calUID]);
-    if (this.title) {
-      lines.push(["SUMMARY", this.title]);
-    }
-    if (this.descriptionText) {
-      lines.push(["DESCRIPTION", this.descriptionText]);
-    }
-    const dateParts = ["VALUE", this.allDay ? "DATE" : "DATE-TIME", "TZID", Intl.DateTimeFormat().resolvedOptions().timeZone];
-    lines.push(["DTSTART", ...dateParts, date2ical(this.startTime, this.allDay)]);
-    lines.push(["DTEND", ...dateParts, date2ical(this.endTime, this.allDay)]);
-    if (this.location) {
-      lines.push(["LOCATION", this.location]);
-    }
-    let organizer = this.participants.find(participant => participant.response == ResponseType.Organizer);
-    if (organizer) {
-      lines.push(["ORGANIZER", "MAILTO:" + organizer.emailAddress]);
-    }
-    if (this.recurrenceRule) {
-      lines.push(this.recurrenceRule.getCalString() + "\r\n");
-    }
-    for (let participant of this.participants) {
-      switch (participant.response) {
-      case ResponseType.Organizer:
-        lines.push(["ATTENDEE", "ROLE", "CHAIR", "PARTSTAT", "ACCEPTED", "CN", participant.name, "MAILTO:" + participant.emailAddress]);
-        break;
-      case ResponseType.Tentative:
-      case ResponseType.Accept:
-      case ResponseType.Decline:
-        lines.push(["ATTENDEE", "PARTSTAT", ParticipationStatus[participant.response], "CN", participant.name, "MAILTO:" + participant.emailAddress]);
-        break;
-      default:
-        lines.push(["ATTENDEE", "RSVP", "TRUE", "CN", participant.name, "MAILTO:" + participant.emailAddress]);
-        break;
-      }
-    }
-    lines.push(["END", "VEVENT"]);
-    lines.push(["END", "VCALENDAR"]);
-    const content = lines.map(line2ical).join("");
-    return { method, content };
-  }
-
   /**
    * Ensures that all recurring instances exist up to the provided date.
    * Must only be called on recurring master events.
@@ -311,42 +256,6 @@ export class Event extends Observable {
       this.calendar.storage.deleteEvent(previous).catch(this.calendar.errorCallback);
     }
   }
-}
-
-function line2ical(line: string | string[]): string {
-  if (typeof line == "string") {
-    return line;
-  }
-  let text = line.shift();
-  let value = line.pop();
-  while (line.length) {
-    text += ";" + line.shift() + "=" + escaped(line.shift(), true);
-  }
-  text += ":" + escaped(value, false);
-  // Lines longer than 75 octets should be folded.
-  // XXX should use TextEncoder to count octets.
-  return text.match(/.{1,75}/gu).join("\r\n ") + "\r\n";
-}
-
-function date2ical(date: Date, allDay: boolean): string {
-  if (!allDay) {
-    return date.toISOString().replace(/-|:|\..../g, "");
-  }
-  return String(date.getFullYear()) + String(date.getMonth() + 1).padStart(2, "0") + String(date.getDate()).padStart(2, "0");
-}
-
-function escaped(s: string, quote: boolean): string {
-  if (quote) {
-    // param-value isn't supposed to include these at all;
-    // maybe we should just delete them?
-    s = s.replace(/["\\]/g, "\\$&").replace(/\r\n?|\n/g, "\\n");
-    if (/[\\:;,]/.test(s)) {
-      s = `"${s}"`;
-    }
-  } else {
-    s = s.replace(/[\\;,]/g, "\\$&").replace(/\r\n?|\n/g, "\\n");
-  }
-  return s;
 }
 
 const k1Day = 86400;
